move `stringDecode()` to `CommonStringExprs` trait
