+++ spark/src/main/spark-3.5/org/apache/comet/shims/CometExprShim.scala
@@ -19,14 +19,14 @@
 package org.apache.comet.shims
 
 import org.apache.comet.expressions.CometEvalMode
+import org.apache.comet.serde.CommonStringExprs
 import org.apache.comet.serde.ExprOuterClass.Expr
-import org.apache.comet.serde.QueryPlanSerde.stringDecode
 import org.apache.spark.sql.catalyst.expressions._
 
 /**
  * `CometExprShim` acts as a shim for for parsing expressions from different Spark versions.
  */
-trait CometExprShim {
+trait CometExprShim extends CommonStringExprs {
     /**
      * Returns a tuple of expressions for the `unhex` function.
      */
