+++ spark/src/main/spark-4.0/org/apache/comet/shims/CometExprShim.scala
@@ -19,8 +19,8 @@
 package org.apache.comet.shims
 
 import org.apache.comet.expressions.CometEvalMode
+import org.apache.comet.serde.CommonStringExprs
 import org.apache.comet.serde.ExprOuterClass.Expr
-import org.apache.comet.serde.QueryPlanSerde.stringDecode
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.expressions.objects.StaticInvoke
 import org.apache.spark.sql.internal.types.StringTypeWithCollation
@@ -29,7 +29,7 @@ import org.apache.spark.sql.types.{BinaryType, BooleanType, StringType}
 /**
  * `CometExprShim` acts as a shim for for parsing expressions from different Spark versions.
  */
-trait CometExprShim {
+trait CometExprShim extends CommonStringExprs {
     /**
      * Returns a tuple of expressions for the `unhex` function.
      */
